+++ source/DSCResources/MSFT_ADDomainTrust/MSFT_ADDomainTrust.psm1
@@ -507,64 +507,6 @@ function Compare-TargetResourceState
 
 <#
     .SYNOPSIS
-        This returns a new object of the type System.DirectoryServices.ActiveDirectory.Domain
-        which is a class that represents an Active Directory Domain Services domain.
-
-    .PARAMETER DirectoryContext
-        The Active Directory context from which the domain object is returned.
-        Calling the Get-ADDirectoryContext gets a value that can be provided in
-        this parameter.
-
-    .NOTES
-        This is a wrapper to enable unit testing of this resource.
-        see issue https://github.com/PowerShell/ActiveDirectoryDsc/issues/324
-        for more information.
-#>
-function Get-ActiveDirectoryDomain
-{
-    [CmdletBinding()]
-    [OutputType([System.DirectoryServices.ActiveDirectory.Domain])]
-    param
-    (
-        [Parameter(Mandatory = $true)]
-        [System.DirectoryServices.ActiveDirectory.DirectoryContext]
-        $DirectoryContext
-    )
-
-    return [System.DirectoryServices.ActiveDirectory.Domain]::GetDomain($DirectoryContext)
-}
-
-<#
-    .SYNOPSIS
-        This returns a new object of the type System.DirectoryServices.ActiveDirectory.Forest
-        which is a class that represents an Active Directory Domain Services forest.
-
-    .PARAMETER DirectoryContext
-        The Active Directory context from which the forest object is returned.
-        Calling the Get-ADDirectoryContext gets a value that can be provided in
-        this parameter.
-
-    .NOTES
-        This is a wrapper to enable unit testing of this resource.
-        see issue https://github.com/PowerShell/ActiveDirectoryDsc/issues/324
-        for more information.
-#>
-function Get-ActiveDirectoryForest
-{
-    [CmdletBinding()]
-    [OutputType([System.DirectoryServices.ActiveDirectory.Forest])]
-    param
-    (
-        [Parameter(Mandatory = $true)]
-        [System.DirectoryServices.ActiveDirectory.DirectoryContext]
-        $DirectoryContext
-    )
-
-    return [System.DirectoryServices.ActiveDirectory.Forest]::GetForest($DirectoryContext)
-}
-
-<#
-    .SYNOPSIS
         This returns the converted value from a Trust Type value to the correct
         Directory Context Type value.